DEV: Changed the way interferogram are displayed (#224)
* FIX: pb with new attributes of LinearCoord and capitalisation of titles.

Titles of axes should be in small capitals for homogeneity along the
program. If we want to capitalise the first letter,
we must do this everywhere!

* DEV: Changed the way interferogram are displayed

Internal scale is in time (1/Hz) but this will not be displayed
to the user. If one needs it to calculate for example apodization
function such as exponential broadening as in NMR,
it can be displayed by setting the parameter _use_time = True
for the x coordinate