MAINT: Refactoring
Analysis is now located out of core.
The reason is that core should concern only what is related to
the structure of data (NDDataset, Project ...) - eventually with closely
related routines such as those in readers or processors.
In analysis are placed routines that make use of these data.

Fitting is actually an analysis process is thus has been placed in Analysis.
Units is now considered as part of core.

This change may break some of the user routines if they import directly
from the analysis sub package for instance. For this reason, any release
concerned by these changes must be 0.3.*.