WIP: refactoring NDMath._op
This function is at the heart of the mathematical calculations on the
data sets and coordinates. However the complexity of its structure made
it very difficult to maintain and debug. Thus, this commit concerns an
important rewrite of this function in order to, at least I hope, make
it more understandable and therefore easier to maintain.