Update test_py_in_docs.py (#388)
* Update test_py_in_docs.py
* FIX: Run docs test only one time on linux and py3.9 and not on windows.
it always fail on win platform.

It is now possible after this fix to run the various *.py files present in docs/userguide and example. 

However this does not work on the GitHub windows.  I don't know much windows so I am not able to fix this problem. BUT I think it is not necessary to test the example on every platform and python version: So I decided to limit to linux/py3.9.