FIX: publish_release (#449)
* FIX: info string

* MAINT: Test following the fix on info string if we can unpin nbsphinx and sphinx

* FIX: :ref: in codespec_adding_reader.rst

* FIX: back to pin versions

I don't now where it is a problem but it doesn't build properly with last versions

* ENH: percent and md used for nbsphinx.

No more need for ipynb files. This suppress the convert steps and speed the building

* DEV: updated colab install pages - add a notebook to launch.

From this commit, *.ipynb are no more ignored. Remove them before
committing if they are not designed to be committed or call them
Untitled.ipynb or scratch*.ipynb

* MAINT: make some file orphan as they are not (yet) used

* FIX: wrong indentation in baselinecorrector string

* FIX: doctring of savgol filter

* FIX: docstring of PFR.fit_to_gas_concentration

* FIX: some issues

* MAINT: Cleanup

* FIX: try to fix sudden error appearing in workflow with jinja2

* DOCS : start refactoring

* FIX: Try to avoid execution of publish_release workflow.