DEV: filter and smooth (#685)
This PR is a revision of the filtering processing methods. 

* Filters has been refactored. A new `Filter` processor class allows to
define various
filters and apply them to a dataset. See `Filtering and Smoothing`
tutorials and `Filter`
for details. Note: Backward compatibility is ensured with the previous
`smooth` and `savgol_filter` methods.

* A `whittaker` filter has been added to the `Filter` processor class.
See `Filtering and Smoothing`
  tutorials and `Filter` for details.

* A `denoise` method based on PCA analysis has been added which allows
to apply a denoising filter to a 2D dataset.

* A `despike` method has been added to the `Filter` processor class.
It allows to remove spikes from a 1D or 2D dataset. This close issues
#688.

Tutorial and examples have been added..